+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Replace terraform kubernetes generic manifests resources with explicit corresponding resources (e.g. daemonsets) because they need the kubeconfig already in plan phase and prevent a single `terraform apply` #87
 - Changed dogu variable in terraform ces-module. All dogus have to be defined. With this change one can change the versions of the necessary dogus like `ldap`.
 - Split terraform google gke example into "bucket" and "ces-cluster" to avoid deleting the bucket all the time. #78
+- Add maintenance window to terraform gke module #92
+- set 1.30 as default k8s version in terraform gke module #92
+- reduce default node disk size in terraform gke module #92
+- add parameters for preemtible and spot VMs in terraform gke module #92
+- (breaking) set `preemtible = false` as default for nodes in terraform gke module #92
 
 ## [v3.0.0] - 2024-06-26